requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.1.4
djangorestframework==3.15.2
ghp-import==2.1.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
watchdog==6.0.0
